replace markdown parser with markdown-it-py
markdown-it-py is Commonmark-compliant, supports various flavors like
gfm, and is based on markdown-it which is well maintained and widely
used, e.g. by StackExchange.

python-markdown does not support Commonmark (deliberately) and insists
on e.g. using four spaces for indenting lists :S

I think Commonmark is the way forward, so using Commonmark parsers is a
good thing :)

Both python-markdown and markdown-it-py don't support math out of the
box and require an extension/plugin to support it. It will probably be a
part of Commonmark at some point:
https://talk.commonmark.org/t/math-rendering-re-visited/4086

Both markdown-it-py and mdit-py-plugins are both available on archlinux.

Note that syntax highlighting requires to include a CSS stylesheet
somewhere. In Anki, common CSS can be shared via a file in the
collection.media folder and putting e.g. `@import url("_anki_base.css");`
in your card template's CSS.

Documentation:

- https://markdown-it-py.readthedocs.io/en/latest/using.html
- https://mdit-py-plugins.readthedocs.io/en/latest